Add StructuredBuffer SRV Test
Vulkan doesn't have the same division between SRV and UAV that DirectX
does, so this refactors the buffer handling code to merge SRVs and UAVs.
We will need different handling for other types of buffers in the future
(i.e. CBVs which map to uniform buffers), but for now this gets pairity
between DX and VK.

The SRV test for structured buffers that @bogner added in his last PR
miscompiles to DXC SPIR-V, so this PR adds a new test to handle more
basic StructuredBuffer SRV use.