Add support for cbuffer resource bindings (#31)
This adds support for cbuffer declarations to the HLSL test suite.

In DirectX, cbuffer declarations map in as Constant Buffer Views (CBV).
In Vulkan, cbuffer declarations map in as uniform buffers.
In Metal, they're just a read-only raw buffer.

There may be some nuances around data layout in constant buffers that aren't consistent between the three platforms so the included test case is extremely simple while we figure out what works and what doesn't.

One important quirk for DirectX is that CBV sizes must be a multiple of 256-bytes, so this implementation pads the remaining bytes with 0's. In the future we may want to instead pad with some garbage since reading out of range gets weird.