Rework Access to Kind (#33)
The mappings between HLSL resource types across APIs is a bit wonky and complicated. It is probably just easier if the test cases define the HLSL resource kind in the YAML and the offloader just maps it to "the right thing"

I have another bigger change I want to make here to separate defining resources from defining their bindings. I'll do that next.