Support shader stages in YAML
This update alters the YAML representation to capture the shader stage and entry function names. Previously I had hacked in `main` in a few places and we surprisingly haven't tripped over it.

I've updated the `TestPipeline` test to use `CSMain` to verify that we can actually change the entry point's name and have everything still work.