Add support for more flexible root signatures (#43)
This adds YAML support for specifying optional `RuntimeSettings` on the `Pipeline` to support runtime-specific configuration information. The first of this data is the `RootParams` list for DirectX to support root constants, descriptors and descriptor tables with arbitrary ordering.

An example usage is:

```yaml
RuntimeSettings:
  DirectX:
    RootParameters:
      - Kind: Constant
        Name: Root
      - Kind: RootDescriptor
        Resource:
            Name: cbuffer
            Kind: ConstantBuffer
      - Kind: DescriptorTable
      - Kind: RootDescriptor
        Resource:
            Name: Out
            Kind: RWStructuredBuffer
```

This would initialize a command buffer for a root signature something like:
`RootConstant(...), CBV(...), DescriptorTable(...), UAV(...)`

---------

Co-authored-by: Farzon Lotfi <farzonl@gmail.com>