Introduce syntax to specify vulkan bindings
This adds a "VKBinding" part for resource descriptors, similar to the already
existing DirectXBinding. These are a bit different from DX in that we only
specify a binding index - the binding can also have a descriptor set, but
that's represented structurally.

There are a few details to be aware of here:

- DXC's behaviour with `register()` annotations is not to differentiate between
  u/t/b types, so I've adjusted quite a few tests to simply avoid bindings that
  would have overlap in vulkan.
- Clang does not support the `vk::binding` attribute or the `-vk-*-shift` flag
  so tricks like the above are the easiest way to make compatible shaders.
- For tests that were buggy on DXC, I've added new tests to cover the shared
  functionality
- I've defaulted the spirv dxc invocations to specify vulkan1.3 to simplify
  command lines