[Vulkan] Use the API version from the device
Initially this code just hard-coded Vulkan 1.1 because you need to set
it before you can query device capabilities (which is messy). This
changes the code to setup a 1.1 device, query it, then recreate the
device with the highest supported version (generally 1.3).

We should update this code to accept a vulkan version from command line
arguments (see #189).