[NFC] Apply clang-tidy (#90)
This applies a set of clang-tidy rules to enforce LLVM coding standards
and some common cleanups.

This PR configures clang-tidy to enforce a swath of rules that comply with and enforce LLVM's coding standards.

Two new CMake options are added:

OFFLOADTEST_USE_CLANG_TIDY - When set to On this runs clang-tidy on non-third-party sources in the offload-test-suite as part of the build process.
OFFLOADTEST_CLANG_TIDY_APPLY_FIX - When set to On passes --fix to clang-tidy to automatically apply fix-its for any issues that have fixes available (Requires OFFLOADTEST_USE_CLANG_TIDY=On).
The enabled warning set is:

clang-diagnostic-* - Clang compiler diagnostics.
llvm-* - LLVM coding standards
misc-* - All miscellaneous rules except:

misc-use-anonymous-namespace, which violates LLVM coding standards which prefers static for functions.
misc-include-cleaner, which did some unholy things to a few source files and caused them to fail to compile.
*misc-non-private-member-variables-in-classes, which conflicts with common patterns in LLVM code.
readability-identifier-naming - Apply LLVM naming conventions

Additionally this PR adds the QualifierAlignment: Left option to the .clang-format configuration because the clang-tidy misc-const-correctness rule adds const to the right of the type which is less common in every codebase ever (Thank's Clang!).