[VK] Fix crash on exit when using VVL
VVL has internal state which conflicts requires a specific cleanup
ordering. Because the VKContext relied on the global dtor, this order
was not respected.

There are 2 solutions:
 - use atexit
 - require an explicit call to vkDestroyDevice.

Thing is, users of `Device` except the object to be usable as a simple
singleton, with no explicit `destroy` to call.
For this reasons, it seems using `atexit` is the cleanest option from
the user perspective.