Rework PR jobs to use a matrix (#231)
This moves the PR workflows into a single matrix configuration which
allows for completely separate workflows from the scheduled jobs. This
fixes two problems with the PR workflows.

First it fixes the README files badges to only be based on scheduled
jobs. This will allow the red/green status badges to reflect the actual
state of main instead of the state of the last PR build.

Second, it uses concurrency groups to cancel in-progress runs for a pull
request if the PR is updated. This should reduce the load on the
builders when they are processing PRs with lots of updates.