Make validation errors a test error.
The current use of the validation layer will emit a message, but it will
not case a test to fail. By adding a call back that returns true for
error messages and warnings, test that have a validation error or
warning will fail.

I've made other changes to keep tests that already pass passing.

xfail test/Basic/TestPipeline.test.

Modify the VkBufferUsageFlagBits for cbuffers. They are uniform buffers
not uniform texel buffers.