Add support for querying extensions (#293)
This adds extensions as a queryable capability through the device
interface for Vulkan. This also gates adding extensions that we had
previously been using based on their availabiltiy.

As a last bit, this does include one small line added in the Changes to
enable MoltenVK on Apple. When targeting Apple platforms the
VK_KHR_PORTABILITY is always enabled. That extension is always required
on Apple platforms so we can hard code that one.