Adjust hyperbolic trig functions to use ULP rules
Unlike the standard trig functions which may be implemented as natural
calculations or as hyperbolic aproximations, the hyperbolic functions
are guaranteed to be hyperbolc aproximations and thus have consistent
precision characteristics.

These can all be guaranteed within 2 ULP, which gives a better match and
will resolve the test failures on AMD GPUs.

The `cosh.16` test is not updated in this PR because it is already using
ULP rules and 2 ULP range specification.