[WARP] Make default WARP version "Last Known Good"
We've encountered a fair number of problems with the System as the
default verison of WARP, instead using a Last Known Good will hopefully
make it easier for contributors to get up and working.

This change adds two new special versions for the `WARP_VERSION` option:
`LKG` and `Latest`. At the moment the `LKG` version is 1.0.14.2, which
is also the latest released version.

This PR also adds a set of build jobs to run periodic builds with the
current WARP pre-release build.

Fixes #340