Adjust hyperbolic trig functions to use ULP rules (#342)
Unlike the standard trig functions which may be implemented as natural
calculations or as hyperbolic aproximations, the hyperbolic functions
are guaranteed to be hyperbolc aproximations and thus have consistent
precision characteristics.

These can all be guaranteed within 2 ULP, which gives a better match and
will resolve the test failures on AMD GPUs.

The `cosh.16` test is not updated in this PR because it is already using
ULP rules and 2 ULP range specification.

I've captured a spec issue to follow up on the very wide range
requirements for `tanh` on NVIDIA GPUs:
https://github.com/microsoft/hlsl-specs/issues/601

This issue has also been observed by WGSL and is reflected in their
spec:
https://github.com/gpuweb/gpuweb/pull/5199

Fixes #326