[NFC] Update flags for portability
This updates all the tests that rely on nan-preservation to pass `-Gis`
which ensures NaN preservation. Any of the tests that are touched in
this PR which also test literal values which could be constant evaluated
also should use `-HV 202x` to have consistent literal types, so that
change is also made to a few test cases.

Fixes #348