Add `all` tests (#360)
Closes #111.

Adds tests for `all` testing 16 bit int types, half, 32 bit types, 64
bit int types, and double.
I moved the bool tests from `all.32.test` to a separate file because
[this bug](https://github.com/llvm/llvm-project/issues/140824) is
causing the vec1 test to fail on Clang, and I didn't want to XFAIL the
rest of the 32 bit tests over it. We can combine them in the future once
the bug is fixed.