Add LIT capabilibies for vulkan extensions (#361)
This adds the ability to conditionalize tests based on available vulkan
extensions. Since a set of our tests depend on float_controls2 when
compiled with Clang, this allows those tests to pass if run on a device
with flaot_controls2 support.

The one additional tweak in this PR is to adjust the dot.fp16 test to
have a wider tolerance range. DXC constant evaluates at a higher
accuracy than the hardware evaluates at and our test assumes constant
folded computation for the expected value. An ULP of 30 is wide enough
to handle both the fp16 "correct" value and the wider computed value
with rounding errors compounded.