Update WARP LKG to 1.0.15
The newly released 1.0.15 fixes a bunch of issues we've encountered in
our testing. This updates the LKG to the new version.

I've tested this locally on arm64 and x86-64, and removed all the
addressed XFAIL annotations. The WaveREadLaneAt.16 test gets a new XFAIL
annotation. This was already failing on our arm64 bots because the issue
occurs in WARP 1.0.14, since upgrading to 1.0.15 fixes more issues than
it introudces I think this is the right tradeoff.

Note: The 1.0.15 release _is not_ based on the 1.0.15-preview release
which caused some problems in our testing. So this should be safe to
adopt.

Fixes #358