Update resource array tests to be more portable
The resource array tests are failing on some of the new hardware we're
bringing up this is for a couple reasons. The main issues stem from
the use of RWBuffer, which may or may not be clamped. This required
wider tolerances for floating point values and caused some integer
values to come back wrong.

This update switches the tests to RWStructuredBuffer instead and adjusts
some of the test checks.

Fixes #335, Fixes #336