Make asuint.64.test compiler agnostic
Amend the test to use literal double values which are encoded to the
same binary represenation by both dxc and clang.  This allows the
expected values to match regardless of which compiler the test
instance is using.

The encoding of the previous arbitrary values used just happened to
differ between dxc and clang, resulting in the test failing when
using dxc as the expected values were those produced by clang.

Fixes #398
Fixes #415
Fixes #418