review: update examples for larger group size
it was noted that if the group size is not larger than the wave size,
the barrier option is redundant

HLSL waves size can be at most 128. Hence, we initialize a group of 512
threads so that it is forced to evaluate over multiple waves.

We also remove the divergent test case as this is not applicable due to
it being undefined behaviour