Add Graphics support to DX, VK and MTL (#441)
This PR adds support for simple graphics pipelines comprised of a vertex
shader and a pixel shader.

The added test case generates an image added in the golden image suite
as https://github.com/llvm/offload-golden-images/pull/1.

Additions to the YAML pipeline description include a new Binding
structure which enables specifying a vertex buffer and associated vertex
attributes as well as a buffer to serve as the render target.

The shader stage enumeration has been extended to support `Vertex` and
`Pixel` stages (`Pixel` mapping to "Fragment" in the MTL and VK APIs).