Add `GroupMemoryBarrierWithGroupSync` tests (#442)
Adds some basic testing of blocking functionality with the
`GroupMemoryBarrierWithGroupSync` intrinsic.

It should be noted that the test cases are sized such that they are
required to be evaluated over multiple waves. In this regard, each
scenario fails using WARP without the barriers. Observing a failure is
dependent on the scheduling of waves.

Resolves: https://github.com/llvm/offload-test-suite/issues/142.