[Vulkan] Remove XFAILs for counter tests (#467)
Counter variable in Clang when targeting Vulkan is now implemented. See
https://github.com/llvm/llvm-project/issues/137032. This removes the
XFAIL from tests that are now passing.

Note that inc_counter_array.test is still failing, but for a different
resson. A new bug was opened to address that issue.

However, a new test is added to test counter arrays.