Add partially mapped resources tests (#476)
This PR is the first step in bringing support to testing with partially
mapped and unmapped resources.
It lays out the infrastructure necessary to initialize resources with
only a certain number of tiles mapped.
The Load overload that takes the 2nd status argument is implemented in
DXC, so we take advantage of that and test that when accessing data in
an unmapped resource, we get 0'd data and the status integer is also 0.

The test is currently unsupported in Clang and Vulkan, but will be
supported in the future.
Fixes https://github.com/llvm/offload-test-suite/issues/182