[Github] Test workflows when modifying them
The postcommit workflows set to run as cron jobs are difficult to test
without landing changes in main. Add a workflow trigger to run them on
pull requests when the workflow definition gets modified.

This will also soon be listed in LLVM's CI best practices document soon.
https://github.com/llvm/llvm-project/pull/172235

Landing this primarily because it makes it easier to test some changes I
want to make soon.