[Github] Hash Pin all External Actions (#619)
This patch hash pins all of the external Github actions that are used
within workflows. We consider this a best practice for LLVM
(https://llvm.org/docs/CIBestPractices.html) as it improves security and
ensures that we control exactly when actions are updated.

The primary motivation for this is so that we can delete setup-windows
from llvm/actions while leaving this repository working until we can
move away from the setup-windows action.